requirements.txt
anyio==4.9.0
apptools==5.3.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
Deprecated==1.2.18
envisage==7.0.3
fabric==3.2.2
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
invoke==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.2
manifold3d==3.0.1
mapbox_earcut==1.0.3
MarkupSafe==2.1.5
matplotlib==3.10.1
mayavi==4.8.3
mpmath==1.3.0
narwhals==1.30.0
networkx==3.3
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
paramiko==3.5.1
paraview-connect==1.0.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.0
pooch==1.8.2
puremagic==1.28
pycollada==0.9
pycparser==2.22
pydicom==3.0.1
pyface==8.0.0
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.1
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-dateutil==2.9.0.post0
pyvista==0.44.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rtree==1.4.0
scipy==1.15.2
scooby==0.10.0
shapely==2.1.0
shiboken6==6.8.2.1
six==1.17.0
sniffio==1.3.1
svg.path==6.3
sympy==1.13.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
traits==7.0.2
traitsui==8.0.0
trimesh==4.6.6
typing_extensions==4.12.2
urllib3==2.3.0
vhacdx==0.0.8.post2
vtk==9.3.1
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
tqdm
